#1854 Refactored tg-toast, separated common part for another toast element for centre refresh on sse event.
