+++ platform-web-ui/src/main/web/ua/com/fielden/platform/web/components/tg-toast.js
@@ -8,51 +8,10 @@ import '/resources/polymer/@polymer/paper-spinner/paper-spinner.js';
 import '/resources/polymer/@polymer/polymer/lib/elements/dom-bind.js';
 
 import { tearDownEvent, containsRestictedTags } from '/resources/reflection/tg-polymer-utils.js';
+import { TgToastBehavior } from '/resources/components/tg-toast-behavior.js';
 
 import { Polymer } from '/resources/polymer/@polymer/polymer/lib/legacy/polymer-fn.js';
 import { html } from '/resources/polymer/@polymer/polymer/lib/utils/html-tag.js';
-
-const paperToastStyle = html`
-    <custom-style>
-        <style>
-            #toast {
-                @apply --layout-horizontal;
-                @apply --layout-center;
-                max-width: 256px;
-                left:0;
-                bottom:0;
-            }
-            .more {
-                padding-left: 8px;
-                color: #03A9F4;
-                font-weight: 800;
-                cursor: pointer;
-            }
-            #toast paper-spinner {
-                width: 1.5em;
-                height: 1.5em;
-                min-width: 1em;
-                min-height: 1em;
-                max-width: 2em;
-                max-height: 2em;
-                padding: 2px;
-                margin-left: 1em;
-                --paper-spinner-layer-1-color: var(--paper-blue-500);
-                --paper-spinner-layer-2-color: var(--paper-blue-500);
-                --paper-spinner-layer-3-color: var(--paper-blue-500);
-                --paper-spinner-layer-4-color: var(--paper-blue-500);
-            }
-            .toast-dialog paper-button {
-                color: var(--paper-light-blue-500);
-                --paper-button-flat-focus-color: var(--paper-light-blue-50);
-            }
-            .toast-dialog paper-button:hover {
-                background: var(--paper-light-blue-50);
-            }
-        </style>
-    </custom-style>`;
-paperToastStyle.setAttribute('style', 'display: none;');
-document.head.appendChild(paperToastStyle.content);
 
 const template = html`
     <paper-toast id="toast" class="paper-toast" text="[[_text]]" on-tap="_showMoreIfPossible" allow-click-through always-on-top duration="0">
@@ -139,17 +98,7 @@ Polymer({
         }
     },
 
-    ready: function () {
-        //Indicates whether toast overlay can be closed via history (back button or not)
-        this.$.toast.skipHistoryAction = true;
-
-        // Styles to truncate the toast text.
-        const label = this.$.toast.$$('#label');
-        label.style.flex = '1';
-        label.style.whiteSpace = 'nowrap';
-        label.style.overflow = 'hidden';
-        label.style.textOverflow = 'ellipsis';
-    },
+    behaviors: [TgToastBehavior],
 
     _hasMoreChanged: function (newValue, oldValue) {
         // let's set a cursor for the whole toast if it can be "clicked"
@@ -218,25 +167,8 @@ Polymer({
         return progress || !hasMore;
     },
 
-    _getPreviousToast() {
-        const toasts = document.querySelectorAll('#toast');
-        let toast = null;
-        let existingToastCount = 0;
-        for (let index = 0; index < toasts.length; index++) {
-            const currToast = toasts.item(index);
-            if (currToast.parentNode === document.body) {
-                existingToastCount++;
-                if (existingToastCount > 1) {
-                    throw 'More than one toast exist in body direct children.';
-                }
-                toast = currToast;
-            }
-        }
-        return toast;
-    },
-
     show: function () {
-        const previousToast = this._getPreviousToast();
+        const previousToast = this.getDocumentToast('toast');
         if (!previousToast) {
             // initial values
             this.$.toast.error = false;
@@ -306,5 +238,9 @@ Polymer({
             this.$.toast._autoCloseCallBack = null;
         }
         this.$.toast.error = false;
+    },
+
+    _toast: function () {
+        return this.$.toast;
     }
 });
